Working myself towards 1.2.4
- Documentation fixes for mixups with column and row vectors
- Fixes to equals/hashcode
- Complex arg()
- serialization for FloatMatrix
- javadoc

+++ src/main/java/org/jblas/ComplexFloatMatrix.java
@@ -1442,7 +1442,7 @@ public class ComplexFloatMatrix {
 		
 		dos.writeInt(data.length);
 		for(int i=0; i < data.length;i++)
-			dos.writeDouble(data[i]);
+			dos.writeFloat(data[i]);
 	}
 	
 	/**
